requirements.txt
pymongo==3.3.0
requests==2.11.1
PyYAML==5.4.1
#ruamel.yaml==0.16.12
#xmltodict==0.10.2
